add package name of encrypt.dal and encrypt.dql
