+++ sharding-core/sharding-core-merge/src/main/java/org/apache/shardingsphere/core/merge/encrypt/dql/EncryptMergeEngine.java
@@ -15,26 +15,28 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.merge.encrypt;
+package org.apache.shardingsphere.core.merge.encrypt.dql;
 
-import com.google.common.base.Optional;
-import org.apache.shardingsphere.spi.encrypt.ShardingEncryptor;
-
-import java.sql.SQLException;
+import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.core.merge.MergeEngine;
+import org.apache.shardingsphere.core.merge.MergedResult;
 
 /**
- * Encryptor meta data.
+ * Encrypt result set merge engine.
  *
  * @author zhangliang
  */
-public interface EncryptorMetaData {
+@RequiredArgsConstructor
+public final class EncryptMergeEngine implements MergeEngine {
     
-    /**
-     * Find encryptor.
-     * 
-     * @param columnIndex column index
-     * @return encryptor
-     * @throws SQLException SQL exception
-     */
-    Optional<ShardingEncryptor> findEncryptor(int columnIndex) throws SQLException;
+    private final EncryptorMetaData metaData;
+    
+    private final MergedResult mergedResult;
+    
+    private final boolean queryWithCipherColumn;
+    
+    @Override
+    public MergedResult merge() {
+        return new EncryptMergedResult(metaData, mergedResult, queryWithCipherColumn);
+    }
 }
