+++ sharding-core/sharding-core-merge/src/test/java/org/apache/shardingsphere/core/merge/encrypt/dal/DescribeTableMergedResultTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.merge.dal.desc;
+package org.apache.shardingsphere.core.merge.encrypt.dal;
 
 import com.google.common.base.Optional;
 import org.apache.shardingsphere.core.execute.sql.execute.result.QueryResult;
